73 - PR comments fix
